Expand CheckingConsoleOutputStream's Functionality
The optional onSeen callback has been changed to onMatchNext. It supplies you with the currently set string for matching and allows you to return a string or to return null. If a string is returned, it continues searching but with the new string you returned instead (return the same string to keep detecting the same thing). If null is returned, that's the signal to stop checking.